backend/requirements.txt
asgiref==3.11.1
decouple==0.0.7
distlib==0.4.0
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.24.2
lxml==6.0.2
pdf2image==1.17.0
pillow==12.1.1
platformdirs==4.9.2
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-docx==1.2.0
python-pptx==1.0.2
sqlparse==0.5.5
typing_extensions==4.15.0
virtualenv==20.37.0
xlsxwriter==3.2.9
